+++ src/DataMapper/Info/Adapter/AbstractAdapter.php
@@ -19,44 +19,227 @@ declare(strict_types=1);
 namespace Phalcon\DataMapper\Info\Adapter;
 
 use Phalcon\DataMapper\Pdo\Connection;
+use Phalcon\DataMapper\Pdo\Exception\Exception;
 
+use function array_column;
+use function array_map;
+use function explode;
 use function in_array;
+use function str_getcsv;
+use function stripos;
 use function strtolower;
 use function strtoupper;
+use function substr;
+use function trim;
 
 /**
- * @phpstan-type ColumnDefinitionSql = array{
- *     _name: string,
- *     _type: string,
- *     _size?: int,
- *     _scale?: int,
- *     _notnull: bool,
- *     _default: mixed,
- *     _autoinc: bool,
- *     _primary: bool,
- *     _options: mixed
- * }
- *
- * @phpstan-type ColumnDefinition = array{
- *     name: string,
- *     type: string,
- *     size: int|null,
- *     scale: int|null,
- *     notnull: bool,
- *     default: mixed,
- *     autoinc: bool,
- *     primary: bool,
- *     options: mixed
- * }
+ * @phpstan-import-type ColumnDefinitionSql from AdapterInterface
+ * @phpstan-import-type ColumnDefinition from AdapterInterface
  */
-abstract class AbstractAdapter
+abstract class AbstractAdapter implements AdapterInterface
 {
+    protected string $currentSchemaSql = '';
+
     /**
      * @param Connection $connection
      */
     public function __construct(
         protected Connection $connection
     ) {
+    }
+
+    /**
+     * Returns the autoincrement or sequence
+     *
+     * @param string $schema
+     * @param string $table
+     *
+     * @return string|null
+     */
+    public function getAutoincSequence(string $schema, string $table): ?string
+    {
+        return null;
+    }
+
+    /**
+     * Return the current schema name
+     *
+     * @return string
+     * @throws Exception
+     */
+    public function getCurrentSchema(): string
+    {
+        /** @var string $currentSchema */
+        $currentSchema = $this->connection->fetchValue($this->currentSchemaSql);
+
+        return $currentSchema;
+    }
+
+    /**
+     * Return the columns in an array with their respective properties
+     *
+     * @param string $schema
+     * @param string $table
+     *
+     * @return array<string, ColumnDefinition>
+     * @throws Exception
+     */
+    public function listColumns(string $schema, string $table): array
+    {
+        $autoInc  = $this->getAutoIncSql();
+        $extended = $this->getExtendedSql();
+
+        $statement = "
+            SELECT
+                c.column_name AS name,
+                c.data_type AS type,
+                COALESCE(
+                    c.character_maximum_length,
+                    c.numeric_precision
+                ) AS size,
+                c.numeric_scale AS numeric_scale,
+                CASE
+                    WHEN c.is_nullable = 'YES' THEN 1
+                    ELSE 0
+                END AS is_nullable,
+                c.column_default AS default_value,
+                $autoInc AS is_auto_increment,
+                CASE
+                    WHEN tc.constraint_type = 'PRIMARY KEY' THEN 1
+                    ELSE 0
+                END AS is_primary,
+                $extended AS extended
+            FROM information_schema.columns c
+                LEFT JOIN information_schema.key_column_usage kcu
+                    ON  c.table_schema = kcu.table_schema
+                    AND c.table_name = kcu.table_name
+                    AND c.column_name = kcu.column_name
+                LEFT JOIN information_schema.table_constraints tc
+                    ON  kcu.table_schema = tc.table_schema
+                    AND kcu.table_name = tc.table_name
+                    AND kcu.constraint_name = tc.constraint_name
+            WHERE c.table_schema = :schema
+            AND   c.table_name = :table
+            ORDER BY c.ordinal_position
+        ";
+
+        /** @var ColumnDefinitionSql[] $columns */
+        $columns = $this->connection->fetchAll(
+            $statement,
+            [
+                'schema' => $schema,
+                'table'  => $table,
+            ]
+        );
+
+        return array_column(
+            array_map(
+                [$this, 'processColumn'],
+                $columns
+            ),
+            null,
+            'name'
+        );
+    }
+
+    /**
+     * Return an array with the schema and table name
+     *
+     * @param string $schemaTable
+     *
+     * @return string[]
+     * @throws Exception
+     */
+    public function listSchemaTable(string $schemaTable): array
+    {
+        $parts = explode('.', $schemaTable, 2);
+
+        if (count($parts) === 1) {
+            return [
+                $this->getCurrentSchema(),
+                $schemaTable,
+            ];
+        }
+
+        return $parts;
+    }
+
+    /**
+     * Returns the SQL for the auto increment column
+     *
+     * @return string
+     */
+    protected function getAutoIncSql(): string
+    {
+        return "''";
+    }
+
+    /**
+     * Returns the actual default value of a column
+     *
+     * @param mixed  $defaultValue
+     * @param string $type
+     *
+     * @return mixed
+     */
+    protected function getDefault(
+        mixed $defaultValue,
+        string $type
+    ): mixed {
+        return $defaultValue;
+    }
+
+    /**
+     * Returns the SQL for the extended field (MySQL)
+     *
+     * @return string
+     */
+    protected function getExtendedSql(): string
+    {
+        return "''";
+    }
+
+    /**
+     * @param ColumnDefinitionSql $column
+     *
+     * @return ColumnDefinition
+     */
+    protected function processColumn(array $column): array
+    {
+        $result = [
+            'name'            => $column['name'],
+            'type'            => $column['type'],
+            'size'            => isset($column['size'])
+                ? (int)$column['size']
+                : null,
+            'scale'           => isset($column['numeric_scale'])
+                ? (int)$column['numeric_scale']
+                : null,
+            'isNullable'      => (bool)$column['is_nullable'],
+            'defaultValue'    => $this->processDefault(
+                $column['default_value'],
+                $column['type']
+            ),
+            'isAutoIncrement' => (bool)$column['is_auto_increment'],
+            'isPrimary'       => (bool)$column['is_primary'],
+            'isUnsigned'      => $this->processSigned(
+                $column['extended'],
+                $column['type']
+            ),
+            'options'         => null,
+        ];
+
+        $extended = trim($column['extended']);
+
+        /**
+         * Enum
+         */
+        if (stripos($extended, 'enum') === 0) {
+            $input             = trim(substr($extended, 4), '()');
+            $result['options'] = str_getcsv($input);
+        }
+
+        return $result;
     }
 
     /**
@@ -71,6 +254,7 @@ abstract class AbstractAdapter
     protected function processDefault(mixed $defaultValue, string $type): mixed
     {
         $type         = strtolower($type);
+        $defaultValue = $this->getDefault($defaultValue, $type);
         $floatTypes   = ['decimal', 'double', 'float', 'numeric', 'real'];
         $keywordTypes = ['CURRENT_DATE', 'CURRENT_TIME', 'CURRENT_TIMESTAMP'];
 
@@ -86,5 +270,16 @@ abstract class AbstractAdapter
             in_array($type, $floatTypes) => (float)$defaultValue,
             default                      => $defaultValue,
         };
+    }
+
+    /**
+     * @param string $value
+     * @param string $type
+     *
+     * @return bool|null
+     */
+    protected function processSigned(string $value, string $type): bool | null
+    {
+        return null;
     }
 }
